invoice_extraction/requirements.txt
requests
streamlit
opencv-python
python-dotenv